MINOR: add missing overrides on TTD store facades (#21759)
This PR adds missing overloads to TDD store facades from StateStore.

Reviewers: Bill Bejeck <bbejeck@apache.org>

+++ streams/src/test/java/org/apache/kafka/streams/state/internals/GenericWindowStoreIteratorFacadeTest.java
@@ -84,7 +84,7 @@ public class GenericWindowStoreIteratorFacadeTest {
 
     @Test
     public void shouldDelegatePeekNextKey() {
-        when(mockedInnerIterator.peekNextKey()).thenReturn(100L, null);
+        when(mockedInnerIterator.peekNextKey()).thenReturn(100L, (Long) null);
 
         assertThat(facade.peekNextKey(), is(100L));
         assertNull(facade.peekNextKey());
